phase3/requirements.txt
# Phase 3 Requirements

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1

# ML / Clustering
umap-learn==0.5.3
hdbscan==0.8.33
scikit-learn==1.3.2
numpy==1.26.2

# LLM
openai==1.3.7

# Database
psycopg2-binary==2.9.9

# Validation
rapidfuzz==3.5.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Phase 2 dependencies (inherited)
redis==5.0.1
celery==5.3.4
